Fix original Path instance modification from __truediv__
Path class's __truediv__ should make a true copy, and not
modify the original instance of Path. Creating a new instance
with `path=self.segments` sets the new instance's path as a
pointer to the old instance's segments.

Also, add test assertion that the original instance is not
modified with __truediv__ when isabsolute is True.

Resolves: #127